+++ game/pkg_Object/Vector2.java
@@ -1,4 +1,4 @@
 package game.pkg_Object;
 
-public record Vector2(int x, int y) {
+public record Vector2(double x, double y) {
 }
